// client/src/types/terrain-type.ts
export type TerrainType =
    | "Wall"
    | "Plain"
    | "Forest"
    | "Thicket"
    | "Bridge"
    | "Stairs"
    | "Floor"
    | "Road"
    | "Pillar"
    | "Lake"
    | "Sea"
    | "River"
    | "Hill"
    | "Mountain"
    | "Cliff"
    | "Barrel"
    | "House"
    | "Fort"
    | "Gate"
    | "Armory"
    | "Vendor"
    | "Village"
    | "Arena"
    | "Chest"
    | "Door"
    | "Ruins"
    | "Throne"; 